Reduce minimum jump height.
The ONLY purpose of this change is to help players understand the controls. Previously, 15% of players didn't work out that jump height is affected by button press duration (when using an xbox 360 controller). Hopefully this will help them see that jump height is variable. If it doesn't, this change should probably be reverted.