Dev: Ents now get a reference to gamestate (gs) on construction, and keep it forever.
Previously I passed in the gs when needed, but passing it around was becoming cumbersome.

Level is now included in the gs, so ents no longer get a separate reference to the level.
This created a circular dependency where gs needs the level, but the level needed a gs to spawn entities. To resolve this, the level no longer automatically spawns entities on construction, so you can create it without a gs then later spawn entities and pass in the gs at the time.